Extract curl command building to it's own module
